+++ src/engine/mg/window.cpp
@@ -57,8 +57,8 @@ bool startFrame() { return !glfwWindowShouldClose(window); }
 float getTime()  { return float(glfwGetTime()); }
 void endFrame() { glfwPollEvents(); }
 
-float getScreenWidth() { return vkContext.screen.width; }
-float getScreenHeight(){ return vkContext.screen.height; };
+float getScreenWidth() { return float(vkContext.screen.width); }
+float getScreenHeight(){ return float(vkContext.screen.height); };
 
 FrameData getFrameData() {
   FrameData frameData = {};
@@ -70,12 +70,12 @@ FrameData getFrameData() {
 
   currentTime = getCurrentTimeUs();
   const auto newDt = (currentTime - prevTime) / 100000.0f;
-  frameData.dt = frameData.dt * 0.99 + 0.01 * newDt;
+  frameData.dt = frameData.dt * 0.99f + 0.01f * newDt;
   prevTime = getCurrentTimeUs();
   const auto newFps = frames / ((currentTime - startTime) / 1000000.0);
   if (frameData.fps == 0)
-    frameData.fps = newFps;
-  frameData.fps = frameData.fps * 0.9 + 0.1 * newFps;
+    frameData.fps = uint32_t(newFps);
+  frameData.fps = uint32_t(frameData.fps * 0.9f + 0.1f * newFps);
 
   int32_t width, height;
   glfwGetWindowSize(window, &width, &height);
@@ -91,7 +91,7 @@ FrameData getFrameData() {
   }
 
   frameData.mouse.prevXY = prevXY;
-  frameData.mouse.xy = cursorPosition(float(width), float(height));
+  frameData.mouse.xy = cursorPosition(width, height);
 
   prevXY = frameData.mouse.xy;
